# practice-app/backend/swagger.yaml
openapi: 3.0.3
info:
  title: Playwright Practice App API
  version: 1.0.0
  description: A mock REST API for practicing Playwright/Selenium API automation. All data is in-memory and resets on server restart or via POST /api/admin/reset.
servers:
  - url: http://localhost:3000/api
paths:
  /auth/login:
    post:
      summary: Login and receive a JWT token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string, example: admin }
                password: { type: string, example: Admin@123 }
      responses:
        '200': { description: OK }
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
  /auth/logout:
    post:
      summary: Logout (stateless)
      responses:
        '200': { description: OK }
  /auth/me:
    get:
      summary: Get current authenticated user
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }
  /users:
    get:
      summary: List users with pagination, filtering, sorting
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: role
          schema: { type: string, enum: [admin, user] }
        - in: query
          name: search
          schema: { type: string }
      responses:
        '200': { description: OK }
    post:
      summary: Create a user (admin only)
      security: [ { bearerAuth: [] } ]
      responses:
        '201': { description: Created }
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
        '403': { description: Forbidden }
  /users/{id}:
    get:
      summary: Get a user by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '404': { description: Not Found }
    put:
      summary: Full update of a user
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: OK }
    patch:
      summary: Partial update of a user
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: OK }
    delete:
      summary: Delete a user (admin only)
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: OK }
        '403': { description: Forbidden }
  /products:
    get:
      summary: List products with filtering, sorting, pagination
      responses:
        '200': { description: OK }
    post:
      summary: Create a product
      security: [ { bearerAuth: [] } ]
      responses:
        '201': { description: Created }
  /products/slow:
    get:
      summary: Simulated slow response endpoint
      parameters:
        - in: query
          name: delay
          schema: { type: integer, example: 3000 }
      responses:
        '200': { description: OK }
  /products/{id}:
    get:
      summary: Get product by id
      responses:
        '200': { description: OK }
        '404': { description: Not Found }
    put:
      summary: Full update
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: OK }
    patch:
      summary: Partial update
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: OK }
    delete:
      summary: Delete product
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: OK }
  /orders:
    get:
      summary: List orders (protected)
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: OK }
    post:
      summary: Create an order (protected)
      security: [ { bearerAuth: [] } ]
      responses:
        '201': { description: Created }
        '400': { description: Bad Request }
  /orders/{id}:
    get:
      summary: Get order by id (protected)
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: OK }
        '404': { description: Not Found }
    delete:
      summary: Delete order (protected)
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: OK }
  /orders/{id}/status:
    patch:
      summary: Update order status (protected)
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: OK }
        '400': { description: Bad Request }
  /files/upload:
    post:
      summary: Upload a single file (multipart/form-data, field "file")
      responses:
        '201': { description: Created }
        '400': { description: Bad Request }
  /files/upload-multiple:
    post:
      summary: Upload multiple files (multipart/form-data, field "files")
      responses:
        '201': { description: Created }
  /files/download/{type}:
    get:
      summary: Download a sample file (txt, csv, json, pdf)
      parameters:
        - in: path
          name: type
          required: true
          schema: { type: string, enum: [txt, csv, json, pdf] }
      responses:
        '200': { description: OK, content: { application/octet-stream: {} } }
        '400': { description: Bad Request }
  /misc/status/{code}:
    get:
      summary: Returns the requested HTTP status code
      parameters:
        - in: path
          name: code
          required: true
          schema: { type: integer, example: 404 }
      responses:
        default: { description: Returns requested status code }
  /misc/delay/{ms}:
    get:
      summary: Simulated network delay (max 15000ms)
      responses:
        '200': { description: OK }
  /misc/nested:
    get:
      summary: Complex nested JSON response
      responses:
        '200': { description: OK }
  /misc/echo:
    post:
      summary: Echoes back the request JSON body
      responses:
        '200': { description: OK }
  /admin/reset:
    post:
      summary: Reset all in-memory data to seed state
      responses:
        '200': { description: OK }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT